Create a superclass for unknown graph.
